Committing naive approach for posterity
So, I tried to do it by iterating over each of the pages, but this is
the dumb way of doing it as you have to add complex logic to keep track
of how far through you are.

Better to work out how many sets of pages (booklets) there will be in
total and then just iterate over each set.

Issue(s): None