Made the main logic a function
Put the main logic around which pages to flip and how to iterate into a
function to make it configurable and take care of the "last booklet"
case.

It's much nicer to work with now I think. Also found places where I had
lots of duplicate code.

Issue(s): None